Gestion de Roles y Permisos

+++ speaksmarter/database/migrations/2024_01_25_016648_create_lessons_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('image_uri', 255)->nullable();
             $table->string('content_uri',255);
             $table->string('pdf_uri',255);
-            $table->unsignedInteger('level_id');
+            $table->unsignedBigInteger('level_id');
             $table->timestamps();
 
             //Declaracion de clave foranea a la tabla referencia levels
